requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn==0.24.0

# Database - Use psycopg2-binary for older Python compatibility
sqlalchemy==2.0.23
psycopg2-binary==2.9.7

# Authentication and Security
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Email and HTTP
jinja2==3.1.2
python-dotenv==1.0.0
aiosmtplib==3.0.1
requests==2.31.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1

# Code Quality (optional)
# black==23.11.0
# ruff==0.1.6
# mypy==1.7.1

# Utilities
pydantic==2.3.0
pydantic-settings==2.0.3
email-validator==2.0.0